requirements.txt
aws-cdk-lib>=2.90.0
constructs>=10.0.0
boto3>=1.26.0
botocore>=1.29.0 